move userMiddleware to shared
