+++ packages/central-server/app/auth/index.js
@@ -4,13 +4,13 @@ import express from 'express';
 import asyncHandler from 'express-async-handler';
 
 import { getPermissions } from '@tamanu/shared/permissions/middleware';
+import { userInfo, userMiddleware } from '@tamanu/shared/auth/userMiddleware';
 
 import { convertFromDbRecord } from '../convertDbRecord';
 import { changePassword } from './changePassword';
 import { resetPassword } from './resetPassword';
 import { login } from './login';
 import { refresh } from './refresh';
-import { userInfo, userMiddleware } from './userMiddleware';
 
 export const DEFAULT_JWT_SECRET = config.auth.secret || crypto.randomUUID();
 export const DEFAULT_JWT_REFRESH_SECRET = config.auth.refreshToken.secret || crypto.randomUUID();
